Tech Debt: Rework Administrative endpoints + Create Service/Repo, Remove Queries (#140)
* Big rework for administrative endpoints + queries
