--- api/src/queries/administrative-activity/index.ts
@@ -1,1 +1,0 @@
-export * from './administrative-activity-queries';
