--- api/src/queries/codes/index.ts
@@ -1,2 +1,0 @@
-export * from './code-queries';
-export * from './db-constant-queries';
