--- api/src/queries/database/index.ts
@@ -1,1 +1,0 @@
-export * from './user-context-queries';
