--- api/src/queries/index.ts
@@ -1,1 +1,0 @@
-export * as Queries from './queries';
