--- api/src/queries/queries.ts
@@ -1,2 +1,0 @@
-export * as administrativeActivity from './administrative-activity';
-export * as database from './database';
