+++ api/src/services/gcnotify-service.ts
@@ -1,6 +1,11 @@
 import axios from 'axios';
-import { ApiError, ApiErrorType } from '../errors/api-error';
+import { ACCESS_REQUEST_ADMIN_EMAIL, ACCESS_REQUEST_APPROVAL_ADMIN_EMAIL } from '../constants/notifications';
+import { IDBConnection } from '../database/db';
+import { ApiError, ApiErrorType, ApiGeneralError } from '../errors/api-error';
 import { IgcNotifyGenericMessage, IgcNotifyPostReturn } from '../interfaces/gcnotify.interface';
+import { AdministrativeRepository } from '../repositories/administrative-repository';
+import { DBService } from './db-service';
+import { KeycloakService } from './keycloak-service';
 
 const EMAIL_TEMPLATE = process.env.GCNOTIFY_ONBOARDING_REQUEST_EMAIL_TEMPLATE || '';
 const SMS_TEMPLATE = process.env.GCNOTIFY_ONBOARDING_REQUEST_SMS_TEMPLATE || '';
@@ -15,7 +20,14 @@ const config = {
   }
 };
 
-export class GCNotifyService {
+export class GCNotifyService extends DBService {
+  administrativeRepository: AdministrativeRepository;
+
+  constructor(connection: IDBConnection) {
+    super(connection);
+
+    this.administrativeRepository = new AdministrativeRepository(connection);
+  }
   /**
    * Send email notification to recipient
    *
@@ -49,7 +61,7 @@ export class GCNotifyService {
   }
 
   /**
-   * Send email notification to recipient
+   * Send sms notification to recipient
    *
    *
    * @param {string} sms
@@ -77,5 +89,78 @@ export class GCNotifyService {
     }
 
     return result;
+  }
+
+  /**
+   * Check if access request is approved, then send Approval email
+   *
+   * @param {number} adminActivityTypeId
+   * @param {string} userIdentifier
+   * @param {string} identitySource
+   * @memberof GCNotifyService
+   */
+  async sendApprovalEmail(adminActivityTypeId: number, userIdentifier: string, identitySource: string) {
+    if (await this.administrativeRepository.checkIfAccessRequestIsApproval(adminActivityTypeId)) {
+      const userEmail = await this.getUserKeycloakEmail(userIdentifier, identitySource);
+      this.sendAccessRequestApprovalEmail(userEmail);
+    }
+  }
+
+  /**
+   * Get users email from username and identity source
+   *
+   * @param {string} userIdentifier
+   * @param {string} identitySource
+   * @return {*}  {Promise<string>}
+   * @memberof GCNotifyService
+   */
+  async getUserKeycloakEmail(userIdentifier: string, identitySource: string): Promise<string> {
+    const keycloakService = new KeycloakService();
+    const userDetails = await keycloakService.getUserByUsername(`${userIdentifier}@${identitySource}`);
+    return userDetails.email;
+  }
+
+  /**
+   * Send Approval Email
+   *
+   * @param {string} userEmail
+   * @memberof GCNotifyService
+   */
+  async sendAccessRequestApprovalEmail(userEmail: string) {
+    const APP_HOST = process.env.APP_HOST;
+    const NODE_ENV = process.env.NODE_ENV;
+
+    const url = `${APP_HOST}/`;
+    const hrefUrl = `[click here.](${url})`;
+    try {
+      await this.sendEmailGCNotification(userEmail, {
+        ...ACCESS_REQUEST_APPROVAL_ADMIN_EMAIL,
+        subject: `${NODE_ENV}: ${ACCESS_REQUEST_APPROVAL_ADMIN_EMAIL.subject}`,
+        body1: `${ACCESS_REQUEST_APPROVAL_ADMIN_EMAIL.body1} ${hrefUrl}`,
+        footer: `${APP_HOST}`
+      });
+    } catch (error) {
+      throw new ApiGeneralError('Failed to send gcNotification approval email', [(error as Error).message]);
+    }
+  }
+
+  /**
+   * Send Email that Access Request has been received
+   *
+   * @memberof GCNotifyService
+   */
+  async sendAccessRequestReceivedEmail() {
+    const ADMIN_EMAIL = process.env.GCNOTIFY_ADMIN_EMAIL || '';
+    const APP_HOST = process.env.APP_HOST;
+    const NODE_ENV = process.env.NODE_ENV;
+
+    const url = `${APP_HOST}/admin/users?authLogin=true`;
+    const hrefUrl = `[click here.](${url})`;
+    this.sendEmailGCNotification(ADMIN_EMAIL, {
+      ...ACCESS_REQUEST_ADMIN_EMAIL,
+      subject: `${NODE_ENV}: ${ACCESS_REQUEST_ADMIN_EMAIL.subject}`,
+      body1: `${ACCESS_REQUEST_ADMIN_EMAIL.body1} ${hrefUrl}`,
+      footer: `${APP_HOST}`
+    });
   }
 }
